Combine test runner and Orchestrator
This patch is a major refactor of the Rust codebase in an effort to
make things a little more organized and minimize code reuse across
testing and actual production environments. Doesn't quite work yet
but the goal is to reproduce the same test runner behavior as before
to start.
