+++ orchestrator/orchestrator/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "orchestrator"
+version = "0.1.0"
+authors = ["Justin Kilpatrick <justin@althea.net>"]
+edition = "2018"
+
+# run the orchestrator by default
+default-run = "orchestrator"
+
+[[bin]]
+name = "orchestrator"
+path = "src/main.rs"
+
+[[bin]]
+name = "test-runner"
+path = "src/test_runner.rs"
+
+[dependencies]
+ethereum_peggy = {path = "../ethereum_peggy"}
+cosmos_peggy = {path = "../cosmos_peggy"}
+peggy_utils = {path = "../peggy_utils"}
+
+#deep_space = {git = "https://github.com/althea-net/deep_space/"}
+#contact = {git = "https://github.com/althea-net/contact"}
+deep_space = {path = "/home/justin/repos/deep_space"}
+contact = {path = "/home/justin/repos/contact"}
+serde_derive = "1.0"
+clarity = "0.3"
+docopt = "1"
+serde = "1.0"
+actix = "0.10"
+actix-web = {version = "3.1", features=["openssl"]}
+actix-rt = "1"
+lazy_static = "1"
+url = "2"
+web30 = "0.7.1"
+num256 = "0.3"
+log = "0.4"
+env_logger = "0.7"
+serde_json = "1.0"
+sha3 = "0.9"
+tokio = "0.2"
+rand = "0.7"
+
+[dev-dependencies]
+actix = "0.10"
