+++ orchestrator/orchestrator/src/tests/mod.rs
@@ -1,0 +1,4 @@
+//! Test runner is a testing script for the Peggy Cosmos module. It is built in Rust rather than python or bash
+//! to maximize code and tooling shared with the validator-daemon and relayer binaries.
+
+mod test_contract_calls;
