+++ orchestrator/peggy_utils/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "peggy_utils"
+version = "0.1.0"
+authors = ["Justin Kilpatrick <justin@althea.net>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+contact = {path = "/home/justin/repos/contact"}
+web30 = "0.7.1"
