--- tests/test-runner/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "test-runner"
-version = "0.1.0"
-authors = ["Justin Kilpatrick <justin@althea.net>"]
-edition = "2018"
-
-[dependencies]
-contact = "0.1.5"
-deep_space = {version = "0.2.1", features = ["peggy", "key_import"]}
-# deep_space = {path = "../../deep_space"}
-# contact = {path = "../../contact"}
-serde_derive = "1.0"
-clarity = "0.3.2"
-docopt = "1"
-serde = "1.0"
-actix-web = {version = "3.0", features=["openssl"]}
-actix-rt = "1"
-lazy_static = "1"
-url = "2"
-web30 = "0.7.1"
-num256 = "0.3"
-rand = "0.7"
-log = "0.4"
-env_logger = "0.7"
-tokio = "0.2"
-sha3 = "0.9"
